% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\docType{class}
\name{Configuration}
\alias{Configuration}
\alias{KHanConfig}
\title{Configuration}
\description{
List of all the configuration settings in this application, which are adjustable by the user.
}
\details{
Configuration Object has no methods. In order to set or get configuration values, use \code{\link{KH.set}} or use \code{\link{KH.get}}
}

\section{Fields}{

\describe{
\item{\strong{defaultFont}:}{character.
Default font used for output for general use; mainly Xetex. \code{defaultFont} defaults to \code{Nazli}.
}
\item{\strong{winFont}:}{character.
Default font used for english text in Xetex. \code{latinFont} defaults to \code{DejaVu Sans}.
}
\item{\strong{latinFont}:}{character.
Default font used for output for MSWord documents; mainly Docx and Xlsx. \code{winFont} defaults to \code{B Nazanin}.
}
\item{\strong{decimal}:}{integer.
Default precission for rounding numbers. \code{decimal} defaults to \code{2}.
}
\item{\strong{workingDirectory}:}{character.
Default working directory for a given script. \code{workingDirectory} defaults to \code{NULL}.
}
\item{\strong{authorName}:}{character.
Name of the package author.
}
\item{\strong{authorEmail}:}{character.
eMail of the package author.
}
}
}

\seealso{
Other Config: \code{\link{KH.get}}, \code{\link{KH.set}}
}
\concept{Config}
