% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{disconnect.default}
\alias{disconnect.default}
\alias{disconnect.Maria}
\alias{disconnect.Oracle}
\alias{disconnect.Postgre}
\alias{disconnect.Sqlite}
\title{Disconnect database resource}
\usage{
\method{disconnect}{default}(obj)
}
\arguments{
\item{obj}{object. One of the four defined database connections.}
}
\value{
boolean. Result of calling database \code{\link[=DBI]{dbDisconnect}}.
}
\description{
Disconnect database resource
}
\details{
Generic function for disconnecting active database resource
}
\seealso{
Other Database: \code{\link{existsTable.default}},
  \code{\link{listTables.default}},
  \code{\link{readAll.default}},
  \code{\link{readSQL.default}},
  \code{\link{readSQLFile.default}},
  \code{\link{sendSQLString.default}},
  \code{\link{writeTable.default}}
}
\concept{Database}
