% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{existsTable.default}
\alias{existsTable.default}
\alias{existsTable.Maria}
\alias{existsTable.Oracle}
\alias{existsTable.Postgre}
\alias{existsTable.Sqlite}
\title{Check Existance of a table}
\usage{
\method{existsTable}{default}(obj, name)
}
\arguments{
\item{obj}{object. One of the four defined database connections.}

\item{name}{character. Name of the table for search.}
}
\value{
boolean.
}
\description{
Check Existance of a table
}
\details{
Checks if a table with the given name is present in connected database.
}
\seealso{
Other Database: \code{\link{disconnect.default}},
  \code{\link{listTables.default}},
  \code{\link{readAll.default}},
  \code{\link{readSQL.default}},
  \code{\link{readSQLFile.default}},
  \code{\link{sendSQLString.default}},
  \code{\link{writeTable.default}}
}
\concept{Database}
