% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wiki_scrape.R
\name{getWikiData}
\alias{getWikiData}
\title{Get JSON Data}
\usage{
getWikiData(url)
}
\arguments{
\item{url}{String containing the Internet location of JSON data.}
}
\description{
Downloads JSON format data from a specified URL. This function is specific to scraping page view data from Wikipedia logs.
Returns a data frame.
}
\examples{
getData(str_URL)
}
\keyword{json,}
\keyword{wikipedia}

