% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wiki_scrape.R
\name{getWikiStats}
\alias{getWikiStats}
\title{Get Wikipedia Stats}
\usage{
getWikiStats(y1, y2, terms)
}
\arguments{
\item{y1}{Beginning Year.}

\item{y2}{Ending Year.}

\item{term}{This is the unique part of the Wikipedia URL representing the article. This may be a vector representing more than one article.}
}
\description{
This function downloads daily page view data for each Wikipedia term passed in.
Data are downloaded from http://stats.grok.se in monthly increments.
Returns a data frame containing Views, Date, and Wikipedia (article titles).
}
\examples{
wiki_data <- getWikiStats(2013, 2014, c("NASA", "Orion_(spacecraft)", "Delta_IV_Heavy", "Exploration_Flight_Test_1"))
}
\keyword{pageviews}
\keyword{wikipedia,}

