% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wiki_scrape.R
\name{getWikiURLs}
\alias{getWikiURLs}
\title{Get Wikipedia URLs}
\usage{
getWikiURLs(y1, y2, term)
}
\arguments{
\item{y1}{Beginning Year.}

\item{y2}{Ending Year.}

\item{term}{This is the unique part of the Wikipedia URL representing the article. This may be a vector representing more than one article.}
}
\description{
This function returns the monthly page view URLs for specific Wikipedia articles stored at http://stats.grok.se
}
\examples{
getWikiURLs(2010, 2014, c("NASA", "Orion_(spacecraft)", "Delta_IV_Heavy", "Exploration_Flight_Test_1"))
}
\keyword{wikipedia}

