% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wiki_scrape.R
\name{google}
\alias{google}
\title{Google}
\usage{
google(term)
}
\arguments{
\item{term}{Search term to find a relevant page.}
}
\description{
Searches the Internet using Google and returns the URL of the first result.
}
\examples{
function_call()
}
\keyword{google,}
\keyword{search}

