% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DbMaria.R
\name{is.Maria}
\alias{is.Maria}
\title{Test relationship between an object and DB object class}
\usage{
is.Maria(obj)
}
\value{
boolean.
}
\description{
Test relationship between an object and DB object class
}
\details{
Check to see if the given object has the class \code{Maria}.
}
\examples{
s1 <- Maria(db_host = '127.0.0.1', db_pass = '123456', db_name = 'test', db_user = 'root')
is.Maria(s1)
#[1] TRUE

s2 <- list(db_host = '127.0.0.1', db_pass = '123456', db_name = 'test', db_user = 'root')
is.Maria(s2)
#[1] FALSE

}
\seealso{
\code{\link{is}}
}
