% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DbSqlite.R
\name{is.Sqlite}
\alias{is.Sqlite}
\title{Test relationship between an object and DB object class}
\usage{
is.Sqlite(obj)
}
\value{
boolean.
}
\description{
Test relationship between an object and DB object class
}
\details{
Check to see if the given object has the class \code{Sqlite}.
}
\examples{
s1 <- Sqlite(db_name = ':memory:')
is.Sqlite(s1)
#[1] TRUE

s2 <- list(db_name = ':memory:')
is.Sqlite(s2)
#[1] FALSE

}
\seealso{
\code{\link{is}}
}
