% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{KH.changeColName}
\alias{KH.changeColName}
\title{Change name of columns in a data set}
\usage{
KH.changeColName(DT, currentTitles, newTitles)
}
\arguments{
\item{DT}{data.table. Two dimension data object which \code{colnames(x)} is meaningful.}

\item{currentTitles}{character. Vector of current column names.}

\item{newTitles}{character. Vector of new names after renaming.}
}
\value{
data.table. The same data object as input with name of columns changed.
}
\description{
Change name of columns in a data set
}
\examples{
{
set.seed(1000)
library(data.table)
d <- data.table(
group = sample(LETTERS[1:3], 10, replace = TRUE),
integer = rnorm(10, 5, 1),
decimal = rnorm(10, 5, 1)
)
colnames(d)
#[1] "group" "integer" "decimal"
d <- KH.changeColName(d, 'integer', 'zeroPrecision')
colnames(d)
#[1] "group" "zeroPrecision" "decimal"
d <- KH.changeColName(d, c('group', 'decimal'), c('newGroup', 'float'))
colnames(d)
#[1] "newGroup" "zeroPrecision" "float"
}
}
\seealso{
Other DataSets: \code{\link{KH.dropColsByName}},
  \code{\link{KH.loadCSV}}
}
\concept{DataSets}
