% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{KH.connect}
\alias{KH.connect}
\title{Connect to a selected database.}
\usage{
KH.connect()
}
\value{
object, an S4 object that inherits from DBIConnection.
This object is used to communicate with the database engine.
}
\description{
Connect to a selected database.
}
\details{
Connect to a database and set \code{\link{db_con}} package configuration to the returned connection object.
There is only one connection object for this package. So only one type of database could be used.
}
\examples{
KH.set('db_name', 'stats_center_4')
KH.set('db_user', 'root')
KH.set('db_pass', '123456')
KH.set('db_host', '127.0.0.1')
KH.connect()
# [1] TRUE
}
\seealso{
\code{\link[=DBI]{dbConnect}} \code{\link{Maria.connect}}

Other Database: \code{\link{KH.disconnect}},
  \code{\link{KH.existsTable}}, \code{\link{KH.readAll}},
  \code{\link{KH.readSQLFile}}, \code{\link{KH.readSQL}},
  \code{\link{KH.sendSQLString}},
  \code{\link{KH.writeTable}}
}
\concept{Database}
