% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{KH.depart}
\alias{KH.depart}
\title{Separate data set into separated output for each department}
\usage{
KH.depart(DT, departmentColumn, outputDir, writeUnivToo = TRUE,
  create.dir = FALSE, output.format = "xlsx", ...)
}
\arguments{
\item{DT}{data.table}

\item{departmentColumn}{character}

\item{outputDir}{string}

\item{writeUnivToo}{logical}

\item{create.dir}{logical}

\item{output.format}{character}

\item{...}{Dots. This is usually \code{...}, although other types are accepted.}
}
\description{
Separate data set into separated output for each department
}
