% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{KH.disconnect}
\alias{KH.disconnect}
\title{Disconnect database resource}
\usage{
KH.disconnect()
}
\value{
boolean. Result of calling database \code{\link[=DBI]{dbDisconnect}}.
}
\description{
Disconnect database resource
}
\details{
Generic function for disconnecting active database rewource.
}
\examples{
KH.disconnect()
# [1] TRUE
}
\seealso{
Other Database: \code{\link{KH.connect}},
  \code{\link{KH.existsTable}}, \code{\link{KH.readAll}},
  \code{\link{KH.readSQLFile}}, \code{\link{KH.readSQL}},
  \code{\link{KH.sendSQLString}},
  \code{\link{KH.writeTable}}
}
\concept{Database}
