% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{KH.dropColsByName}
\alias{KH.dropColsByName}
\title{Drop column or columns from a data.table by using the names.}
\usage{
KH.dropColsByName(DT, currentTitles)
}
\arguments{
\item{DT}{data.table Two dimension data object which \code{colnames(x)} is meaningful.}

\item{currentTitles}{character. Vector of current column names.}
}
\value{
data.table. The same object as input with columns removed.
}
\description{
Drop column or columns from a data.table by using the names.
}
\examples{
set.seed(1000)
library(data.table)
d <- data.table(
group = sample(LETTERS[1:3], 10, replace = TRUE),
integer = rnorm(10, 5, 1),
integer1 = rnorm(10, 5, 1),
integer2 = rnorm(10, 5, 1),
decimal = rnorm(10, 5, 1)
)
colnames(d)
#[1] "group" "integer" "integer1" "integer2" "decimal"
d <- KH.dropColsByName(d, 'integer')
colnames(d)
#[1] "group" "integer1" "integer2" "decimal"

d <- KH.dropColsByName(d, c('integer1', 'integer2'))
colnames(d)
#[1] "group" "decimal"
}
\seealso{
Other DataSets: \code{\link{KH.changeColName}},
  \code{\link{KH.loadCSV}}, \code{\link{KH.loadXlsx}}
}
\concept{DataSets}
