% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{KH.existsTable}
\alias{KH.existsTable}
\title{Check Existance of a table and optionally check if it is empty}
\usage{
KH.existsTable(collectionName, populated = FALSE)
}
\arguments{
\item{collectionName}{string. Name of table}

\item{populated}{boolean. Check for records.}
}
\value{
boolean.
}
\description{
Check Existance of a table and optionally check if it is empty
}
\details{
Check for existence of th given table. If it is found and contains any records.
If \code{populated} is set to \code{TRUE}, and the table is empty, the result would be \code{FALSE}.
}
\seealso{
Other Database: \code{\link{KH.connect}},
  \code{\link{KH.disconnect}}, \code{\link{KH.readAll}},
  \code{\link{KH.readSQLFile}}, \code{\link{KH.readSQL}},
  \code{\link{KH.sendSQLString}},
  \code{\link{KH.writeTable}}
}
\concept{Database}
