% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{KH.get}
\alias{KH.get}
\title{Get module configurations.}
\usage{
KH.get(variable, default = NULL)
}
\arguments{
\item{variable}{String. Name of the variable to search for.}

\item{default}{Mixed. Default value if the object is not present, defaults to \code{NULL}.}
}
\value{
Mixed. Based on the requested variable.
}
\description{
Get module configurations.
}
\details{
Critical package-wide variables are saved in a module environment.
This function reads a given variable from this environment only.
}
\examples{
KH.get('defaultFont')
# "Nazli"
KH.get('notPresentVariable', default = NA)
# NA

KH.set('defaultFont', 'B Nazanin')
KH.get('defaultFont')
# "B Nazanin"
}
\seealso{
Other Config: \code{\link{Configuration}},
  \code{\link{KH.set}}
}
\concept{Config}
\keyword{internals}
