% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{KH.ifNull}
\alias{KH.ifNull}
\title{Assign value to the variable.}
\usage{
KH.ifNull(userSet, defaultSet)
}
\arguments{
\item{userSet}{Mixed The desired value to assign as output.}

\item{defaultSet}{Mixed. The default used when the \code{userSet} is \code{NULL}.}
}
\value{
Mixed. New value of the variable.
}
\description{
Assign value to the variable.
}
\details{
Useful for times checking for \code{NULL} and decide the outcome of a value.
}
\examples{
KH.ifNull(NULL, 'B Nazanin')
#[1] "B Nazanin"

KH.ifNull('defaultFont', 'B Nazanin')
#[1] "defaultFont"

}
\seealso{
Other Utils: \code{\link{KH.changeDecimal}},
  \code{\link{KH.roundAll}}
}
\concept{Utils}
