% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{KH.IQR}
\alias{KH.IQR}
\title{Get \code{IQR} with rounded output}
\usage{
KH.IQR(x, decimal = KH.get(variable = "decimal", default = 1), ...)
}
\arguments{
\item{x}{numeric. Vector of numbers to calculate interquartile range.}

\item{decimal}{integer. Set precision for vectors in resultant interquartile range. Default is \code{decimal} in package.}

\item{...}{Dots. An arbitrary number and variety of arguments, passing arguments on to \code{IQR} function.}
}
\value{
numeric. IQR of the vector rounded to the given precision.
}
\description{
Get \code{IQR} with rounded output
}
\details{
Just a wrapper for the generic \code{IQR} function with extra rounding.
}
\examples{
set.seed(100)
KH.IQR(c(rnorm(10, sd = 1)))
#[1] 0.74
set.seed(100)
KH.IQR(c(rnorm(10, sd = 1)), decimal = 3)
#[1] 0.738
}
\seealso{
\code{\link{IQR}}

Other Stats: \code{\link{KH.max}}, \code{\link{KH.mean}},
  \code{\link{KH.median}}, \code{\link{KH.min}},
  \code{\link{KH.nineNumbers.by}},
  \code{\link{KH.nineNumbers.table}},
  \code{\link{KH.nineNumbers.total}},
  \code{\link{KH.quant.10}}, \code{\link{KH.quant.25}},
  \code{\link{KH.quant.75}}, \code{\link{KH.quant.n}},
  \code{\link{KH.sd}}
}
\concept{Stats}
