% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{KH.loadCSV}
\alias{KH.loadCSV}
\title{Load data from a CSV file into `data.table`}
\usage{
KH.loadCSV(csv_filename, key = NULL, index = NULL, colClasses = NULL)
}
\arguments{
\item{csv_filename}{string. Full URL to the CSV file.}

\item{key}{atomic or vector of string. Name of fields used in the keys. \emph{\strong{NOT Used for now.}}}

\item{index}{atomic or vecotr of string. Name of fields used in indexing data. \emph{\strong{NOT Used for now.}}}

\item{colClasses}{list of named strings. Name and classes of fields used to read and load data. \emph{\strong{NOT Used for now.}}}
}
\value{
data.table. An object of type \code{data.table}
}
\description{
Load data from a CSV file into `data.table`
}
\details{
Use only data of \code{data.table} type. Many functions of this package work with this type of data only.
}
\note{
All the functions in this package accept data.table only.
}
\examples{
{
input_file <- system.file("extdata", "input_file.csv", package = "KHanSUtils")
KH.loadCSV(input_file)

#year month  day dep_delay arr_delay carrier origin dest air_time distance hour
#1: 2014  TRUE TRUE        14        13      AA    JFK  LAX      359     2475    9
#2: 2014  TRUE TRUE        -3        13      AA    JFK  LAX      363     2475   11
#3: 2014  TRUE TRUE         2         9      AA    JFK  LAX      351     2475   19
#4: 2014  TRUE TRUE        -8       -26      AA    LGA  PBI      157     1035    7
#5: 2014  TRUE TRUE         2         1      AA    JFK  LAX      350     2475   13
#6: 2014  TRUE TRUE         4         0      AA    EWR  LAX      339     2454   18
}
}
\concept{DataLoaders}
