% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.R
\name{KH.loadXlsx}
\alias{KH.loadXlsx}
\title{Read an Existing Excel Workbook}
\usage{
KH.loadXlsx(xls_filename, sheetID = 1, key = NULL, index = NULL, ...)
}
\arguments{
\item{xls_filename}{string. Full URL to the xlsx file.}

\item{sheetID}{string|number. Name or index of the sheet to read from the opened file. Default is 1.}

\item{key}{atomic or vector of string. Name of fields used in the keys. \emph{\strong{NOT Used for now.}}}

\item{index}{atomic or vecotr of string. Name of fields used in indexing data. \emph{\strong{NOT Used for now.}}}

\item{...}{Dots. An arbitrary number and variety of arguments, passing arguments on
to \code{\link[=openxlsx]{read.xlsx}}}
}
\value{
data.table. An object of type \code{data.table}
}
\description{
Read an Existing Excel Workbook
}
\details{
Read an existing Excel workbook into data frame.
}
\note{
All the functions in this package accept data.table only.
}
\examples{
fileName = system.file("extdata", "input_file.xlsx", package = "KHanSUtils")
sheetID = 1

KH.loadXlsx(fileName, sheetID)
#   year month  day dep_delay arr_delay carrier origin dest air_time distance hour
# 1: 2014     1   1        14        13      AA    JFK  LAX      359     2475    9
# 2: 2014    10  31         1       -30      UA    LGA  IAH      201     1416   14
# 3: 2014     1   1        -3        13      AA    JFK  LAX      363     2475   11
# 4: 2014    10  31        -5       -14      UA    EWR  IAH      189     1400    8
# 5: 2014     1   1         2         9      AA    JFK  LAX      351     2475   19
# 6: 2014    10  31        -8        16      MQ    LGA  RDU       83      431   11
# 7: 2014     1   1        -8       -26      AA    LGA  PBI      157     1035    7
# 8: 2014    10  31        -4        15      MQ    LGA  DTW       75      502   11
# 9: 2014     1   1         2         1      AA    JFK  LAX      350     2475   13
#10: 2014    10  31        -5         1      MQ    LGA  SDF      110      659    8
#11: 2014     1   1         4         0      AA    EWR  LAX      339     2454   18
}
\seealso{
Other DataSets: \code{\link{KH.changeColName}},
  \code{\link{KH.dropColsByName}}, \code{\link{KH.loadCSV}}
}
\concept{DataSets}
