% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{KH.max}
\alias{KH.max}
\title{Get \code{max} with rounded output}
\usage{
KH.max(x, decimal = KH.get(variable = "decimal", default = 1), ...)
}
\arguments{
\item{x}{numeric. Vector of numbers to calculate max.}

\item{decimal}{integer. Set precision for vectors in resultant max. Default is \code{decimal} in package.}

\item{...}{Dots. An arbitrary number and variety of arguments, passing arguments on to \code{max} function.}
}
\value{
numeric. Max of the vector rounded to the given precision.
}
\description{
Get \code{max} with rounded output
}
\details{
Just a wrapper for the generic \code{max} function with extra rounding.
}
\examples{
set.seed(100)
KH.max(c(rnorm(10, sd = 1)))
#[1] -0.02
set.seed(100)
KH.max(c(rnorm(10, sd = 1)), decimal = 3)
#[1] -0.018
}
\seealso{
\code{\link{max}}

Other Stats: \code{\link{KH.IQR}}, \code{\link{KH.mean}},
  \code{\link{KH.median}}, \code{\link{KH.min}},
  \code{\link{KH.nineNumbers.by}},
  \code{\link{KH.nineNumbers.table}},
  \code{\link{KH.nineNumbers.total}},
  \code{\link{KH.quant.10}}, \code{\link{KH.quant.25}},
  \code{\link{KH.quant.75}}, \code{\link{KH.quant.n}},
  \code{\link{KH.sd}}
}
\concept{Stats}
