% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{KH.mean}
\alias{KH.mean}
\title{Get \code{mean} with rounded output}
\usage{
KH.mean(x, decimal = KH.get(variable = "decimal", default = 1), ...)
}
\arguments{
\item{x}{numeric. Vector of numbers to calculate mean.}

\item{decimal}{integer. Set precision for vectors in resultant mean. Default is \code{decimal} in package.}

\item{...}{Dots. An arbitrary number and variety of arguments, passing arguments on to \code{mean} function.}
}
\value{
numeric. Mean of the vector rounded to the given precision.
}
\description{
Get \code{mean} with rounded output
}
\details{
Just a wrapper for the generic \code{mean} function with extra rounding.
}
\examples{
set.seed(100)
KH.mean(c(rnorm(10, sd = 1)))
#[1] -0.02
set.seed(100)
KH.mean(c(rnorm(10, sd = 1)), decimal = 3)
#[1] -0.018
}
\seealso{
\code{\link{mean}}

Other Stats: \code{\link{KH.IQR}}, \code{\link{KH.max}},
  \code{\link{KH.median}}, \code{\link{KH.min}},
  \code{\link{KH.nineNumbers.by}},
  \code{\link{KH.nineNumbers.table}},
  \code{\link{KH.nineNumbers.total}},
  \code{\link{KH.quant.10}}, \code{\link{KH.quant.25}},
  \code{\link{KH.quant.75}}, \code{\link{KH.quant.n}},
  \code{\link{KH.sd}}
}
\concept{Stats}
