% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{KH.nineNumbers.by}
\alias{KH.nineNumbers.by}
\title{Produce table of nine-numbers grouped by a given column in the data.}
\usage{
KH.nineNumbers.by(DT, x, groupBy, titles = NULL, decimal = "default",
  order = NULL)
}
\arguments{
\item{DT}{data.table date.table of at least one column of numbers and at least one column of factoring element to calculate statistics.}

\item{x}{string. Name of vector containing the data to extract statistics for.}

\item{groupBy}{string. Name of vector containig grouping factors.}

\item{titles}{string. Vector of strings with 9 member as col.names of the results.}

\item{decimal}{integer. Set precision for vectors in resultant quantile. Default is \code{decimal} in package.}

\item{order}{string. Not implemented yet. Sort order of columns.}
}
\description{
Produce table of nine-numbers grouped by a given column in the data.
}
\details{
For an appropriate insight into the data 9 numbers are consulted usually.
These nine numbers consist of \code{length, minimum, lower-hinge, median, upper-hinge, maximum, mean, sd}
of a data vector. This function creates a row for each grouped item.
}
\examples{
{
set.seed(100)
library(data.table)
df <- data.table(
   x = rnorm(20),
   y = sample(LETTERS[1:3], 20, replace = TRUE)
)
x=KH.nineNumbers.by(df, 'x', 'y')
}
}
\seealso{
Other Stats: \code{\link{KH.IQR}}, \code{\link{KH.max}},
  \code{\link{KH.mean}}, \code{\link{KH.median}},
  \code{\link{KH.min}}, \code{\link{KH.nineNumbers.table}},
  \code{\link{KH.nineNumbers.total}},
  \code{\link{KH.quant.10}}, \code{\link{KH.quant.25}},
  \code{\link{KH.quant.75}}, \code{\link{KH.quant.n}},
  \code{\link{KH.sd}}
}
\concept{Stats}
