% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{KH.nineNumbers.table}
\alias{KH.nineNumbers.table}
\title{Produce complete table of nine-numbers grouped by a given column in the data plus a total row.}
\usage{
KH.nineNumbers.table(DT, x, groupBy, titles = NULL, rowname = NULL,
  decimal = "default", order = "default")
}
\arguments{
\item{DT}{data.table. data.table of at least one column of numbers and at least one column of factoring element to calculate statistics.}

\item{x}{string. Name of vector containing the data to extract statistics for.}

\item{groupBy}{string. Vector containig name of grouping factors columns.}

\item{titles}{string. Vector of strings with 9 member as col.names of the results.}

\item{rowname}{string. A title for the row. This is handy for attaching the output to the table from \code{Stats.nineNumbers.by}.}

\item{decimal}{integer. Set precision for vectors in resultant quantile. Default is \code{decimal} in package.}

\item{order}{string. Not implemented yet. Sort order of columns.}
}
\description{
Produce complete table of nine-numbers grouped by a given column in the data plus a total row.
}
\details{
For an appropriate insight into the data 9 numbers are consulted usually.
These nine numbers consist of \code{length, minimum, lower-hinge, median, upper-hinge, maximum, mean, sd}
of a data vector. This function creates both a table and a row for total data.
}
\examples{
{
set.seed(100)
df <- data.table(
   x = rnorm(20),
   y = sample(LETTERS[1:3], 20, replace = TRUE)
)
x=KH.nineNumbers.by(df, 'x', 'y')
z=KH.nineNumbers.total(df, 'x')
y=KH.nineNumbers.table(df, 'x', 'y')
}
}
\seealso{
Other Stats: \code{\link{KH.IQR}}, \code{\link{KH.max}},
  \code{\link{KH.mean}}, \code{\link{KH.median}},
  \code{\link{KH.min}}, \code{\link{KH.nineNumbers.by}},
  \code{\link{KH.nineNumbers.total}},
  \code{\link{KH.quant.10}}, \code{\link{KH.quant.25}},
  \code{\link{KH.quant.75}}, \code{\link{KH.quant.n}},
  \code{\link{KH.sd}}
}
\concept{Stats}
