% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{KH.quant.10}
\alias{KH.quant.10}
\title{Get \code{quantile} with rounded output for probability of 0.10}
\usage{
KH.quant.10(x, decimal = KH.get(variable = "decimal", default = 1), ...)
}
\arguments{
\item{x}{numeric. Vector of numbers to calculate quantile.}

\item{decimal}{integer. Set precision for vectors in resultant quantile. Default is \code{decimal} in package.}

\item{...}{Dots. An arbitrary number and variety of arguments, passing arguments on to \code{quantile} function.}
}
\value{
numeric. quantile of the vector rounded to the given precision.
}
\description{
Get \code{quantile} with rounded output for probability of 0.10
}
\examples{
set.seed(100)
KH.quant.10(c(rnorm(10, sd = 1)))
#  10\%
#-0.61
set.seed(100)
KH.quant.10(c(rnorm(10, sd = 1)), decimal = 3)
#  10\%
#-0.606
}
\seealso{
Other Stats: \code{\link{KH.IQR}}, \code{\link{KH.max}},
  \code{\link{KH.mean}}, \code{\link{KH.median}},
  \code{\link{KH.min}}, \code{\link{KH.nineNumbers.by}},
  \code{\link{KH.nineNumbers.table}},
  \code{\link{KH.nineNumbers.total}},
  \code{\link{KH.quant.25}}, \code{\link{KH.quant.75}},
  \code{\link{KH.quant.n}}, \code{\link{KH.sd}}
}
\concept{Stats}
