% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{KH.readAll}
\alias{KH.readAll}
\title{Read all the rows and columns in the given table.}
\usage{
KH.readAll(collectionName)
}
\arguments{
\item{collectionName}{String. Name of a table in the database.}
}
\value{
data.table Curser result set from server.
For example see \code{\link{Db.readSQL}}
}
\description{
Read all the rows and columns in the given table.
}
\details{
#' Read all of rows in the given table containing all the columns in the result set.
}
\seealso{
\code{\link{KH.readSQL}}

Other Database: \code{\link{KH.connect}},
  \code{\link{KH.disconnect}},
  \code{\link{KH.existsTable}},
  \code{\link{KH.readSQLFile}}, \code{\link{KH.readSQL}},
  \code{\link{KH.sendSQLString}},
  \code{\link{KH.writeTable}}
}
\concept{Database}
