% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{KH.readSQL}
\alias{KH.readSQL}
\title{Read sequel from string and return the data}
\usage{
KH.readSQL(sqlString, key = NULL)
}
\arguments{
\item{sqlString}{string. Complete SQL string.}

\item{key}{Optional key column(s).}
}
\value{
data.table. Curser result set from server.
}
\description{
Read sequel from string and return the data
}
\details{
Send SQL to the databse server and read the data as a data.table. Presumably the SQL string is
quoted in MariaDB standards (i.e.: using back quotes).
}
\examples{
KH.connect()
KH.readSQL('SELECT * FROM oci_departments LIMIT 1;', key = c('department_id'))
# department_id              title short_name campus ruling_department
# 1            10 معارف وعلوم انسانی      معارف   سایر                 0
KH.disconnect()
# [1] TRUE
}
\seealso{
Other Database: \code{\link{KH.connect}},
  \code{\link{KH.disconnect}},
  \code{\link{KH.existsTable}}, \code{\link{KH.readAll}},
  \code{\link{KH.readSQLFile}},
  \code{\link{KH.sendSQLString}},
  \code{\link{KH.writeTable}}
}
\concept{Database}
