% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{KH.readSQLFile}
\alias{KH.readSQLFile}
\title{Read sequel from a file and return the data}
\usage{
KH.readSQLFile(sqlFile)
}
\arguments{
\item{sqlFile}{string. Full name of a text file containg a valid SQL query.}
}
\value{
data.table Curser result set from server.
For example see \code{\link{Db.readSQL}}
}
\description{
Read sequel from a file and return the data
}
\details{
Send SQL saved in a file to the databse server and read the data as a data.table.
Presumably the SQL string is quoted in MariaDB standards (i.e.: using back quotes).
For example see \code{\link{KH.readSQL}}
}
\seealso{
\code{\link{KH.readSQL}}

Other Database: \code{\link{KH.connect}},
  \code{\link{KH.disconnect}},
  \code{\link{KH.existsTable}}, \code{\link{KH.readAll}},
  \code{\link{KH.readSQL}}, \code{\link{KH.sendSQLString}},
  \code{\link{KH.writeTable}}
}
\concept{Database}
