% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{KH.sd}
\alias{KH.sd}
\title{Get \code{sd} with rounded output}
\usage{
KH.sd(x, decimal = KH.get(variable = "decimal", default = 1), ...)
}
\arguments{
\item{x}{numeric. Vector of numbers to calculate standard deviation.}

\item{decimal}{integer. Set precision for resultant standard deviation. Default is \code{decimal} in package.}

\item{...}{Dots. An arbitrary number and variety of arguments, passing arguments on to \code{sd} function.}
}
\value{
numeric. Standard deviation of the vector rounded to the given precision.
}
\description{
Get \code{sd} with rounded output
}
\details{
Just a wrapper for the generic \code{sd} function with extra rounding.
}
\examples{
set.seed(100)
KH.sd(c(rnorm(10, mean = 0)))
#[1] 0.56
set.seed(100)
KH.sd(c(rnorm(10, mean = 0)), decimal = 3)
#[1] 1.561
}
\seealso{
\code{\link{sd}}

Other Stats: \code{\link{KH.IQR}}, \code{\link{KH.max}},
  \code{\link{KH.mean}}, \code{\link{KH.median}},
  \code{\link{KH.min}}, \code{\link{KH.nineNumbers.by}},
  \code{\link{KH.nineNumbers.table}},
  \code{\link{KH.nineNumbers.total}},
  \code{\link{KH.quant.10}}, \code{\link{KH.quant.25}},
  \code{\link{KH.quant.75}}, \code{\link{KH.quant.n}}
}
\concept{Stats}
