% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{KH.sendSQLString}
\alias{KH.sendSQLString}
\title{Send sequel from a string}
\usage{
KH.sendSQLString(sqlString)
}
\arguments{
\item{sqlString}{string. Complete SQL string.}
}
\value{
boolean. Result of execution.
}
\description{
Send sequel from a string
}
\details{
Execute queries without result sets on the server.
}
\examples{
\dontrun{
KH.sendSQLString(paste0('TRUNCATE TABLE `a_table`;'))
}
}
\seealso{
Other Database: \code{\link{KH.connect}},
  \code{\link{KH.disconnect}},
  \code{\link{KH.existsTable}}, \code{\link{KH.readAll}},
  \code{\link{KH.readSQLFile}}, \code{\link{KH.readSQL}},
  \code{\link{KH.writeTable}}
}
\concept{Database}
