% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{KH.writeTable}
\alias{KH.writeTable}
\title{Write and append data to the table.}
\usage{
KH.writeTable(DT, tableName, field.types = NULL)
}
\arguments{
\item{DT}{data.table Dataset ready to save in the database.}

\item{tableName}{string. Name of table.}

\item{field.types}{list. Named list of field types}
}
\value{
boolean. Result of saving.
}
\description{
Write and append data to the table.
}
\details{
Write a data.table to the table in chunks. Chunking the data frame is done for avoiding database
queue overflow.
}
\seealso{
Other Database: \code{\link{KH.connect}},
  \code{\link{KH.disconnect}},
  \code{\link{KH.existsTable}}, \code{\link{KH.readAll}},
  \code{\link{KH.readSQLFile}}, \code{\link{KH.readSQL}},
  \code{\link{KH.sendSQLString}}
}
\concept{Database}
