% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHanSUtils.R
\docType{package}
\name{KHanSUtils}
\alias{KHanSUtils}
\alias{KHanSUtils-package}
\title{KHanSUtils: KHan R Programming Utilities}
\description{
A series of files containing everyday-used methods and data structures.
    This is the true successor of the KHanPgrad package. This is a more general package, 
    which does not contain any domain specific tools.
}
\details{
The only function you're likely to need from roxygen2 is [roxygenize()].
Otherwise refer to the vignettes to see how to format the documentation.
}
\note{
DT is used to denote using data.table as the data set
}
\keyword{internal}
