% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{listTables.default}
\alias{listTables.default}
\alias{listTables.Maria}
\alias{listTables.Oracle}
\alias{listTables.Postgre}
\alias{listTables.Sqlite}
\title{List tables in the database}
\usage{
\method{listTables}{default}(obj)
}
\arguments{
\item{obj}{object. One of the four defined database connections.}
}
\value{
boolean. Result of calling database \code{\link[=DBI]{dblistTables}}.
}
\description{
List tables in the database
}
\details{
Generic function for getting list of tables in active database resource
}
\seealso{
Other Database: \code{\link{disconnect.default}},
  \code{\link{existsTable.default}},
  \code{\link{readAll.default}},
  \code{\link{readSQL.default}},
  \code{\link{readSQLFile.default}},
  \code{\link{sendSQLString.default}},
  \code{\link{writeTable.default}}
}
\concept{Database}
