% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprs.R
\name{makeContrasts2}
\alias{makeContrasts2}
\title{Make contrast DESeq2-style}
\usage{
makeContrasts2(contrast, levels)
}
\arguments{
\item{contrast}{vector of three elements (factor, level2, level1)}
}
\value{
contrast table for level2-level1 comparison
}
\description{
Make contrast DESeq2-style
}
\examples{
es.design <- model.matrix(~0+condition, data=pData(es.norm))
fit <- lmFit(es.norm, es.design)
fit2 <- contrasts.fit(fit, makeContrasts2(c("condition", "Treatment", "Control"), 
                                         levels=es.design))
fit2 <- eBayes(fit2)
de <- topTable(fit2, adjust.method="BH", number=Inf)
}
