% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DbMaria.R
\name{Maria}
\alias{Maria}
\title{Create a database connection object using MariaDB DBMS.}
\usage{
Maria(db_name, db_user, db_pass, db_host, db_port = 3306)
}
\arguments{
\item{db_name}{string. Name of the database to load.}

\item{db_user}{string. User identification for the DBMS.}

\item{db_pass}{string. Password identifying the user.}

\item{db_host}{string. Name of the server hosting the database.}

\item{db_port}{integer. Port the server is listening.}
}
\value{
Maria, an S3 object of type Maria containing the required parameter to act as
an interface to RMariaDB.
}
\description{
Create a database connection object using MariaDB DBMS.
}
\details{
Connect to a database and set \code{\link{con_obj}} in the object itself.
}
\examples{
s1 <- Maria(db_host = '127.0.0.1', db_name = 'test', db_user = 'root', db_pass = '123456')
s1
#$db_host
#[1] "127.0.0.1"

#$db_port
#[1] 3306

#$db_name
#[1] "test"

#$db_user
#[1] "root"

$db_pass
#[1] "123456"

#$con_obj
#<MariaDBConnection>
#  Host:    127.0.0.1
#  Server:  5.5.60-MariaDB
#  Client:  5.5.60-MariaDB

#attr(,"class")
#[1] "Maria"

}
\seealso{
\code{\link[=DBI]{dbConnect}} \code{\link{Database}}
}
