% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistate.R
\name{normalizeExpressions}
\alias{normalizeExpressions}
\title{Normalize expression table}
\usage{
normalizeExpressions(exprs, zero.rm = T, log2 = T, quantile = T)
}
\arguments{
\item{exprs}{Table with expressions}

\item{zero.rm}{If TRUE removes genes with zero expression in all samples}

\item{log2}{It TRUE applies log2 transform. Zeroes are replaced with minimal non-zero element for sample}

\item{quantile}{If TRUE applies quantile normalization}
}
\description{
Normalize expression table
}

