% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DbOracle.R
\name{Oci}
\alias{Oci}
\title{Create a database connection object using Oracle DBMS.}
\usage{
Oci(db_name, db_user, db_pass, db_host, db_port = 1521)
}
\arguments{
\item{db_name}{string. Name of the database to load.}

\item{db_user}{string. User identification for the DBMS.}

\item{db_pass}{string. Password identifying the user.}

\item{db_host}{string. Name of the server hosting the database.}

\item{db_port}{integer. Port the server is listening.}
}
\value{
Oci, an S3 object of type Oci containing the required parameter to act as
an interface to ROracle.
}
\description{
Create a database connection object using Oracle DBMS.
}
\details{
Connect to a database and set \code{\link{con_obj}} in the object itself.
}
\note{
Oci connection is specially design to connect to AUT pGrad Portal. It will not function
well in other environments.
}
\examples{
con <- Oci(db_host = 'localhost', db_name = 'test', db_user = 'root', db_pass = '123456')
con
#$db_host
#[1] "localhost"

#$db_port
#[1] "1521"

#$db_name
#[1] "test"

#$db_user
#[1] "root"

#$db_pass
#[1] "123456"

#$con_str
#[1] "(DESCRIPTION=(ADDRESS_LIST =(ADDRESS=(PROTOCOL=TCP)(HOST=localhost)(PORT=1521)))(CONNECT_DATA=(SID=test)))"

#$con_obj
#User name:             root
#Connect string:        (DESCRIPTION=(ADDRESS_LIST =(ADDRESS=(PROTOCOL=TCP)(HOST=localhost)(PORT=1521)))(CONNECT_DATA=(SID=test)))
#Server version:        11.2.0.4.0
#Server type:           Oracle RDBMS
#Results processed:     0
#OCI prefetch:          FALSE
#Bulk read:             1000
#Bulk write:            1000
#Statement cache size:  0
#Open results:          0

#attr(,"class")
#[1] "Oci"

}
\seealso{
\code{\link[=DBI]{dbConnect}} \code{\link{Database}}
}
