% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Plots.createBoxPlot}
\alias{Plots.createBoxPlot}
\title{Draw a horizontal boxplot.}
\usage{
Plots.createBoxPlot(dataSet, x, y, title, xlab = NULL, ylab = NULL,
  notch = FALSE)
}
\arguments{
\item{dataSet}{data frame. Data to put in the plot.}

\item{x}{string. Name of column used as factoring elemnt and shown in the vertical axis.}

\item{y}{string. Name of column used for building the plot. This should be a numeric vector column.}

\item{title}{string. Main title of the graph.}

\item{xlab}{string. Label of the x-axis, which is shown vertically here.}

\item{ylab}{string. Label of the y-axis, which is shown horizontally here.}

\item{notch}{boolean. If notch is TRUE, a notch is drawn in each side of the boxes.
If the notches of two plots do not overlap this is ‘strong evidence’ that the two medians differ.
\code{\link{boxplot}} for more information and \code{\link[=boxplot]{stats}} for the calculations used.}
}
\value{
Class of type ggplot. Use \code{\link{plot}} for really draw the plots.
}
\description{
Draw a horizontal boxplot.
}
\details{
Create horizontal boxplots based on given value and partitions by the x column.
}
\note{
In a horizontal plot, x and y axes are swapped. So x-axis is vertical and y-axis is horizontal.
}
\examples{
\dontrun{
Plots.createBoxPlot(sbm.data, 'student_status', 'average_semester', 'توزیع میانگین ترمی دانشجویان',)
}
}
\concept{Plots}
