% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DbPostgre.R
\name{Postgre}
\alias{Postgre}
\title{Create a database connection object using PostgreSQL DBMS.}
\usage{
Postgre(db_name, db_user, db_pass, db_host, db_port = 5432)
}
\arguments{
\item{db_name}{string. Name of the database to load.}

\item{db_user}{string. User identification for the DBMS.}

\item{db_pass}{string. Password identifying the user.}

\item{db_host}{string. Name of the server hosting the database.}

\item{db_port}{integer. Port the server is listening.}
}
\value{
Postgre, an S3 object of type Postgre containing the required parameter to act as
an interface to PostgreSQL.
}
\description{
Create a database connection object using PostgreSQL DBMS.
}
\details{
Connect to a database and set \code{\link{con_obj}} in the object itself.
}
\examples{
con <- Postgre(db_host = 'localhost', db_name = 'test', db_user = 'root', db_pass = '123456')
con
#$db_host
#[1] "127.0.0.1"

#$db_port
#[1] 5432

#$db_name
#[1] "test"

#$db_user
#[1] "keyhan"

#$db_pass
#[1] "123456"

#$con_obj
#<PostgreSQLConnection>

#attr(,"class")
#[1] "Postgre"

}
\seealso{
\code{\link[=DBI]{dbConnect}} \code{\link{Database}}
}
