% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{readAll.default}
\alias{readAll.default}
\title{Read all the rows and columns in the given table.}
\usage{
\method{readAll}{default}(obj, tableName)
}
\arguments{
\item{obj}{object. One of the four defined database connections.}

\item{tableName}{String. Name of a table in the database.}
}
\value{
data.table Curser result set from server.
}
\description{
Read all the rows and columns in the given table.
}
\details{
Read all of rows in the given table containing all the columns in the result set.
}
\seealso{
Other Database: \code{\link{disconnect.default}},
  \code{\link{existsTable.default}},
  \code{\link{listTables.default}},
  \code{\link{readSQL.default}},
  \code{\link{readSQLFile.default}},
  \code{\link{sendSQLString.default}},
  \code{\link{writeTable.default}}
}
\concept{Database}
