% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{readSQL.default}
\alias{readSQL.default}
\title{Read sequel from string and return the data}
\usage{
\method{readSQL}{default}(obj, sqlString, key)
}
\arguments{
\item{obj}{object. One of the four defined database connections.}

\item{sqlString}{string. Complete SQL string.}

\item{key}{Optional key column(s).}
}
\value{
data.table. Curser result set from server.
}
\description{
Read sequel from string and return the data
}
\details{
Send SQL to the databse server and read the data as a data.table.
}
\seealso{
Other Database: \code{\link{disconnect.default}},
  \code{\link{existsTable.default}},
  \code{\link{listTables.default}},
  \code{\link{readAll.default}},
  \code{\link{readSQLFile.default}},
  \code{\link{sendSQLString.default}},
  \code{\link{writeTable.default}}
}
\concept{Database}
