% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{readSQLFile.default}
\alias{readSQLFile.default}
\title{Read sequel from a file and return the data}
\usage{
\method{readSQLFile}{default}(obj, sqlFile)
}
\arguments{
\item{sqlFile}{string. Full name of a text file containg a valid SQL query.}
}
\value{
data.table Curser result set from server.
}
\description{
Read sequel from a file and return the data
}
\details{
Send SQL saved in a file to the databse server and read the data as a data.table.
}
\seealso{
Other Database: \code{\link{disconnect.default}},
  \code{\link{existsTable.default}},
  \code{\link{listTables.default}},
  \code{\link{readAll.default}},
  \code{\link{readSQL.default}},
  \code{\link{sendSQLString.default}},
  \code{\link{writeTable.default}}
}
\concept{Database}
