% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{removeLayer}
\alias{removeLayer}
\title{Remove layer from ggplot object
from https://stackoverflow.com/questions/13407236/remove-a-layer-from-a-ggplot2-chart}
\usage{
removeLayer(ggplot2_object, geom = NULL, color = NULL)
}
\description{
Remove layer from ggplot object
from https://stackoverflow.com/questions/13407236/remove-a-layer-from-a-ggplot2-chart
}
