% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wiki_scrape.R
\name{right}
\alias{right}
\title{Right}
\usage{
right(str, n)
}
\arguments{
\item{str}{String value to subset.}

\item{n}{Number of characters to return from right side of string.}
}
\description{
Returns the right n characters of a specified string.
}
\examples{
right('Hello World', 5)
}
\keyword{left,}
\keyword{right,}
\keyword{subset}

