% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{sendSQLString.default}
\alias{sendSQLString.default}
\title{Send sequel from a string}
\usage{
\method{sendSQLString}{default}(obj, sqlString)
}
\arguments{
\item{obj}{object. One of the four defined database connections.}

\item{sqlString}{string. Complete SQL string.}
}
\value{
boolean. Result of execution.
}
\description{
Send sequel from a string
}
\details{
Execute queries without result sets on the server.
}
\seealso{
Other Database: \code{\link{disconnect.default}},
  \code{\link{existsTable.default}},
  \code{\link{listTables.default}},
  \code{\link{readAll.default}},
  \code{\link{readSQL.default}},
  \code{\link{readSQLFile.default}},
  \code{\link{writeTable.default}}
}
\concept{Database}
