% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DbSqlite.R
\name{Sqlite}
\alias{Sqlite}
\title{Create a database connection object using Sqlite}
\usage{
Sqlite(db_name)
}
\arguments{
\item{db_name}{string. Path to the file containing the database to load.
\code{:memory:} creates a temporary in-memory database.}
}
\value{
Sqlite, an S3 object of type Sqlite containing the required parameter to act as
an interface to RSQLite.
}
\description{
Create a database connection object using Sqlite
}
\details{
Connect to a database and set \code{\link{con_obj}} in the object itself.
}
\examples{
con <- Sqlite(db_name = ':memory:')
con
#$db_name
#[1] ":memory:"

#$con_obj
#<SQLiteConnection>
#  Path: :memory:
#  Extensions: TRUE

#attr(,"class")
#[1] "Sqlite"

}
\seealso{
\code{\link[=DBI]{dbConnect}} \code{\link{Database}}
}
