% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEL-Samples-N-Notes.R
\encoding{UTF-8}
\name{test_function}
\alias{test_function}
\title{Some dummy function to list selected ROxygen documentation tags}
\usage{
test_function(DT, ...)
}
\arguments{
\item{DT}{data.table}

\item{...}{Dots. This is usually \code{...}, although other types are accepted.
linear-time algorithms to test chordality of graphs, test acyclicity
of hypergraphs, and selectively reduce acyclic hypergraphs.
*SIAM Journal of Computation* **13**, 566-579.}
}
\value{
data.table
}
\description{
Some dummy function to list selected ROxygen documentation tags
}
\section{Warning}{

You must not call this function unless ...
\code{\link[magrittr]{\%>\%}}
\emph{italics} \strong{bold}
\code{r_function_call(with = "arguments")}: r_function_call(with = "arguments")
\preformatted{Ther
IS
<strong>text</strong>
HERE
}
\subsection{Exceptions}{
   Apart from the following special cases...
}
See `::is_falsy` for the definition of what is _falsy_
and what is _truthy_.
}

\examples{
{
input_file <- system.file("extdata", "input_file.csv", package = "KHanSUtils")
DT <- KH.loadCSV(input_file)
}
}
\references{
Robert E Tarjan and Mihalis Yannakakis. (1984). Simple
}
\keyword{internals}
