% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wiki_scrape.R
\name{WikiPVPlot}
\alias{WikiPVPlot}
\title{Wikipedia Plot}
\usage{
WikiPVPlot(input)
}
\arguments{
\item{input}{Wikipedia page view data frame from getWikiStats() function. Columns requires are Views, Date, and Wikipedia (article titles).}
}
\description{
Plots Wikipedia page view data over time.
}
\examples{
WikiPVPlot(wiki_data)
}
\keyword{plot}
\keyword{wikipedia,}

