% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wiki_scrape.R
\name{wikiURL}
\alias{wikiURL}
\title{Wikipedia URL}
\usage{
wikiURL(term)
}
\arguments{
\item{term}{Search term to find a relevant Wikipedia page.}
}
\description{
This function searches the internet using Google for the search term and the word 'Wikipedia'. Returns the URL of the first result.
}
\examples{
wikiURL('Orion capsule')
}
\keyword{google,}
\keyword{search}
\keyword{wikipedia,}

