% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database.R
\name{writeTable.default}
\alias{writeTable.default}
\alias{writeTable.Maria}
\alias{writeTable.Oracle}
\alias{writeTable.Postgre}
\alias{writeTable.Sqlite}
\title{Write and append data to the table.}
\usage{
\method{writeTable}{default}(obj, DT, tableName)
}
\arguments{
\item{obj}{object. One of the four defined database connections.}

\item{DT}{data.table Dataset ready to save in the database.}

\item{tableName}{string. Name of table.}
}
\value{
boolean. Result of saving.
}
\description{
Write and append data to the table.
}
\details{
Write a data.table to the table in chunks. Chunking the data frame is done for avoiding database
queue overflow.
}
\seealso{
Other Database: \code{\link{disconnect.default}},
  \code{\link{existsTable.default}},
  \code{\link{listTables.default}},
  \code{\link{readAll.default}},
  \code{\link{readSQL.default}},
  \code{\link{readSQLFile.default}},
  \code{\link{sendSQLString.default}}
}
\concept{Database}
