% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\encoding{UTF-8}
\name{KH.changeDecimal}
\alias{KH.changeDecimal}
\title{Format numbers in the table so they are correct in Persian.}
\usage{
KH.changeDecimal(DT, columns = c())
}
\arguments{
\item{DT}{data.table. Numerical data required for changing the decimal point.}

\item{columns}{string|number. Columns' names or numbers in the data to change.}
}
\description{
Format numbers in the table so they are correct in Persian.
}
\details{
Decimal point in persian is different from that in English. This function replaces it so the
decimal numbers are read correctly
}
\examples{
{
KH.changeDecimal(1.23)
#[1] "1٫23"
KH.changeDecimal(12.3)
#[1] "12٫3"
KH.changeDecimal('a.23')
#[1] "a.23"

columns <- c('dep_x','arr_x')
input_file <- system.file("extdata", "input_file.csv", package = "KHanSUtils")
DT <- KH.loadCSV(input_file)

DT[, c('dep_x','arr_x') := list(dep_delay/10, arr_delay/10)]
DT
#   dep_delay arr_delay carrier origin dest air_time distance hour dep_x arr_x
#1:        14        13      AA    JFK  LAX      359     2475    9   1.4   1.3
#2:        -3        13      AA    JFK  LAX      363     2475   11  -0.3   1.3
#3:         2         9      AA    JFK  LAX      351     2475   19   0.2   0.9
#4:        -8       -26      AA    LGA  PBI      157     1035    7  -0.8  -2.6
#5:         2         1      AA    JFK  LAX      350     2475   13   0.2   0.1
#6:         4         0      AA    EWR  LAX      339     2454   18   0.4   0.0

KH.changeDecimal(DT, columns)
#   dep_delay arr_delay carrier origin dest air_time distance hour dep_x arr_x
#1:        14        13      AA    JFK  LAX      359     2475    9   1٫4   1٫3
#2:        -3        13      AA    JFK  LAX      363     2475   11  -0.3   1٫3
#3:         2         9      AA    JFK  LAX      351     2475   19   0٫2   0٫9
#4:        -8       -26      AA    LGA  PBI      157     1035    7  -0.8  -2.6
#5:         2         1      AA    JFK  LAX      350     2475   13   0٫2   0٫1
#6:         4         0      AA    EWR  LAX      339     2454   18   0٫4     0
}
}
\seealso{
Other Utils: \code{\link{KH.get}}, \code{\link{KH.ifNull}},
  \code{\link{KH.set}}
}
\concept{Utils}
\keyword{internals}
