% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{KH.set}
\alias{KH.set}
\title{Save a new value for module environment.}
\usage{
KH.set(variable, value)
}
\arguments{
\item{variable}{string. Name of the variable or function to change.}

\item{value}{mixed. New value of the object.}
}
\value{
mixed. New value of the variable.
}
\description{
Save a new value for module environment.
}
\details{
In order to change a default value or function definition, set the value using this method.
It stops if a foreign key is requested
}
\note{
Use this only to redefine a variable. It could not be used to create a variable.
}
\examples{
KH.set('defaultFont', 'B Nazanin')
#[1] "B Nazanin"

KH.set('decimal', 3)
#[1] 3
\dontrun{
KH.set('new_var', 2)

KH.set('hello', function(){print('hello')})
# function(){print("hello")}
}
}
\seealso{
Other Config: \code{\link{Configuration}},
  \code{\link{KH.get}}
}
\concept{Config}
