% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.gtf}
\alias{read.gtf}
\title{Reads gtf file to data.table}
\usage{
read.gtf(file, attrs.to.extract = c("gene_id", "transcript_id", "gene_type",
  "gene_name"), features.to.extract = NULL)
}
\arguments{
\item{file}{Path to file}
}
\description{
Reads gtf file to data.table
}
